refactor cache
