+++ src/utils.rs
@@ -17,7 +17,7 @@ use std::fs::File;
 use std::io::BufReader;
 use std::path::{Path, PathBuf};
 use std::thread;
-use std::time::{Duration, Instant};
+use std::time::{Duration};
 
 use exr::prelude as exrs;
 use exr::prelude::*;
@@ -34,6 +34,7 @@ use std::sync::mpsc::{Receiver, Sender};
 use strum::Display;
 use strum_macros::EnumIter;
 
+use crate::cache::Cache;
 use crate::image_editing::EditState;
 use crate::settings::PersistentSettings;
 
@@ -129,48 +130,7 @@ impl ExtendedImageInfo {
     }
 }
 
-#[derive(Debug)]
 
-pub struct Cache {
-    data: HashMap<PathBuf, CachedImage>,
-    pub cache_size: usize,
-}
-
-#[derive(Debug)]
-pub struct CachedImage {
-    data: RgbaImage,
-    created: Instant,
-}
-
-impl Cache {
-    fn get(&self, path: &Path) -> Option<RgbaImage> {
-        self.data.get(path).map(|c| c.data.clone())
-    }
-
-    pub fn insert(&mut self, path: &Path, img: RgbaImage) {
-        self.data.insert(
-            path.into(),
-            CachedImage {
-                data: img,
-                created: std::time::Instant::now(),
-            },
-        );
-        if self.data.len() > self.cache_size {
-            debug!("Cache limit hit, deleting oldest");
-            let mut latest = std::time::Instant::now();
-            let mut key = PathBuf::new();
-
-            for (p, c) in &self.data {
-                if c.created < latest {
-                    latest = c.created;
-                    key = p.clone();
-                }
-            }
-
-            _ = self.data.remove(&key);
-        }
-    }
-}
 
 #[derive(Debug)]
 pub struct Player {
